requirements.txt
anthropic==0.40.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
